requirements.txt
lib4sbom >= 0.7.2
lib4package >= 0.2.0
packageurl-python
packaging
